RubyConf India 2014 in Goa
This was my first talk in front of more than 50 people. And it turned
out to be about 400. Wow.

I would have loved to extended my stay, because India is beautiful.
Unfortunately university called :unamused:
Next time I'm going to stay longer.

And, as the common theme in most of my travels: Food. Wow. Amazing.

(I also spent about 8 hours in total in Delhi, but this was at the
airport, won't count that :wink:)