Roadtrip through Scotland
Three friends of mine and I went on a roadtrip through Scotland. That
was amazing. We met up in Edinburgh, spent a few days there. This city
is so beautiful, I can't believe it.

We continued on to Grantown on Spey. This little city is in the middle
of both the Cairngorms National Park and the whisky region Speyside. So
there was a lot of beautiful landscapes and a lot of really amazing
whiskys. And the B&B was so good – oh my. This breakfast will haunt me
in my dreams. Trips included Loch Ness and a lot of other nice
landmarks and historical places.

After another 5 days in the Speyside, we continued our journey to
Aberfoyle in the middle of Trossachs National Park.

This was a lot to see for just 10 days. Wow.