RedDotRubyConf in Singapore
Gave my third talk at a Ruby conference this year, it was really nice.

![RedDotRuby](http://www.reddotrubyconf.com/assets/logo-vertical-8bf5613cc59ab2a2aa763c8719a00b33.png)

The conference was really nice, a lot of amazing people there.

The country itself is interesting. On the one hand it is very beautiful
and full of excellent food and really, really nice people.
On the other hand it is scary as hell. When you enter the country, you
will be informed that you will be sentenced to death if you ever have
drugs with you. I'm not a drug baron, still I was a little scared. Later
I read more on the wiki page and it scared the living crap out of me.
But everything turned out fine :wink:

If you go there, definitely go to the [Gardens by the Bay]
](https://en.wikipedia.org/wiki/Gardens_by_the_Bay). Really awesome.

Movies watched on the flight:

* Star Trek: The future begins
* Man of Tai Chi
* 12 Years a Slave
* Mandela

Books read on the flight:

* No Place to Hide
* Efficiency in Development Workflows (from Codeship)