Been to Nürnberg for Developer Week
This conference was a bit more business-y than the conferences I usually
go to. They also had a strong focus on the Microsoft stack and – to my
surprise – Delphi. On the first evening I was in the middle of a bunch
of Delphi developers. Didn't expect that to happen.

But I think it was a good thing to go to a conference so entirely
different from the conferences I usually go to. There were a lot of nice
people and getting their perspective on where the industry is right now
and what our challenges are was more than interesting.

Nürnberg itself is a really beautiful, tiny city. At least the historic
core of the city is. And of course I had to try the sausages there. They
were really, really good. If you ever go there:

1. Go from the main station to the castle, you will go through most of
the historical part of the city which is really beautiful.
2. Sit down in one of the places that says something about 'Würstchen' and
get 'Blaue Zipfel'
3. Enjoy :)