In November 2014 I've been to Poznan for Polyconf
I went there with @nerdbabe, @bascht, @railsbros-dirk and @madziaf.
We had a great time, a lot of fun and really, really good food :smile:

I also gave a talk. Almost nobody was there, even the two main
organizers were sound asleep :wink: But it was great to spend a lot of
time with the people mentioned above :blue_heart: