Walking the Speyside with @13abylon
We walked the [Speyside
Way](http://www.walkhighlands.co.uk/speyside-way.shtml) from Buckie to
Grantown-on-Spey. We also were in Edinburgh and briefly in Aberdeen.

We had our tent with us which we slept in. We had everything from great
weather to pouring rain, from drinking excellent Whisky to dragging
ourselves through swamp and cow shit. We slept in our tent, but also in
some nice BnBs along the way.

Oh, and of course Haggis and full scottish breakfast. I miss both.

It was super exhausting and super amazing :smile: Thanks for joining me,
@13abylon.