Chaos Communication Camp 2015
This was my first Chaos Communication Camp. An event I wanted to go to
for a long, long time. It was even better than I imagined. I can't even
describe how good it was :smile:

Thanks to the amazing village and the great people. There were too many
there to mention, but they all made it a very special occassion.

Going home from there was really hard. It really was like we had build
our own wonderful world there in the middle of nowhere.

Special shout out to two friends that literally came halfway across the
world for this event: @bkerley and @elcuervo :heart: