innoQ Far-Far-Away Event in Milan
This was the year I started working for innoQ, an amazing company.
Part of working there is going to six so called events a year. One is
far, far away -- and this event was in Milan.

We went to World Expo and in general had a good time there :smile: