Bodo in Scandinavia
I traveled a bit. I visited Copenhagen (which is beautiful) and took a
train to Stockholm (which I didn't like that much). Due to a train
problem I ended up in Malmö for 2 hours (which is also beautiful from
what I could see).

One little note for Stockholm: If you ever plan to visit a usergroup
there, don't show up in a nerd shirt. You will be the only one and feel
underdressed. Yes, even on a JS usergroup :wink:.